# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'EduCBT API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: "All endpoints (except authentication & onboarding) require a Sanctum Bearer token.\n\n**How to authenticate:**\n1. Call `POST /api/auth/login` with your `identifier` and `password`.\n2. The response will include an `access_token`.\n3. Pass this token as a Bearer token in the `Authorization` header for all subsequent requests.\n\n**For tenants (schools):** Use the tenant's subdomain URL and a tenant user's credentials (admin, teacher, or student).\n**For super admins:** Use the central domain URL and super admin credentials."
security:
  -
    default: []
paths:
...
  '/api/class-levels/{classLevelId}/arms/{id}':
    patch:
      summary: 'Update a class arm name.'
      operationId: updateAClassArmName
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Classes & Arms'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Arm name.'
                  example: null
              required:
                - name
    delete:
      summary: 'Delete a class arm (must have no students assigned).'
      operationId: deleteAClassArmmustHaveNoStudentsAssigned
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Classes & Arms'
    parameters:
      -
        in: path
        name: classLevelId
        description: 'The class level UUID.'
        example: id
        required: true
        schema:
          type: string
      -
        in: path
        name: id
        description: 'The arm UUID.'
        example: id
        required: true
        schema:
          type: string
...
  '/api/super-admin/plans/{id}':
    get:
      summary: "Get a single plan's details."
      operationId: getASinglePlansDetails
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Super Admin: Billing & Subscriptions'
    put:
      summary: 'Update a subscription plan.'
      operationId: updateASubscriptionPlan
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Super Admin: Billing & Subscriptions'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Plan name.'
                  example: null
                max_students:
                  type: integer
                  description: 'Max students.'
                  example: null
                max_teachers:
                  type: integer
                  description: 'Max teachers.'
                  example: null
                max_exams_per_term:
                  type: integer
                  description: 'Max exams per term.'
                  example: null
                price_monthly:
                  type: integer
                  description: 'Monthly price.'
                  example: null
                price_yearly:
                  type: integer
                  description: 'Yearly price.'
                  example: null
                features:
                  type: array
                  description: 'Features list.'
                  example: null
                  items:
                    type: string
                    nullable: true
    delete:
      summary: 'Delete a subscription plan.'
      operationId: deleteASubscriptionPlan
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Super Admin: Billing & Subscriptions'
    parameters:
      -
        in: path
        name: id
        description: 'The plan UUID.'
        example: id
        required: true
        schema:
          type: string
...
  '/api/super-admin/tenants/{id}':
    get:
      summary: 'Get a single tenant with its domains.'
      operationId: getASingleTenantWithItsDomains
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Super Admin: Tenant Management'
    put:
      summary: 'Update a tenant.'
      operationId: updateATenant
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Super Admin: Tenant Management'
    delete:
      summary: 'Permanently delete a tenant.'
      operationId: permanentlyDeleteATenant
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Super Admin: Tenant Management'
    parameters:
      -
        in: path
        name: id
        description: 'The tenant UUID.'
        example: id
        required: true
        schema:
          type: string